pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "books-rec-api"
version = "0.1.0"
description = "Recommendation Service MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.46",
  "psycopg[binary]>=3.3.3",
  "alembic>=1.18.4",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "mypy>=1.17.0",
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/books_rec_api", "scripts", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_unused_ignores = true
mypy_path = "src"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
